backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Smart debt tracking with two‑party confirmation, reminders, Trust Score and QR profiles"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "pydantic-settings>=2.14.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.26",
    "supabase>=2.29.0",
    "uvicorn>=0.46.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project.scripts]
main = "app.main:main"
server = "app.main:main"


[tool.mypy]
check_untyped_defs = true
explicit_package_bases = true
warn_unreachable = true
warn_redundant_casts = true

# Performance optimizations
cache_fine_grained = true
incremental = true
sqlite_cache = true
follow_imports = "skip"

[tool.ruff]
line-length = 150
target-version = "py312"

[tool.ruff.lint]
select = [
    "F",    # Pyflakes
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
]
ignore = [
    "E501",  # Long lines
    "E731"   # lambda vs. def
]

[tool.vulture]
paths = ["app"]
exclude = [".tmp"]
min_confidence = 80

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.5",
]
